I have made Simulation an aggregate root that has Environment because I feel like the environment cannot exist without the simulation and there is no need for outside objects to reference the environment.
