+++ simulation.py
@@ -6,7 +6,6 @@ import matplotlib.pyplot as plt
 from matplotlib.animation import FuncAnimation
 
 from config import Configuration, config_error
-from environment import build_hospital
 from infection import find_nearby, infect, recover_or_die, compute_mortality,\
 healthcare_infection_correction
 from motion import update_positions, out_of_bounds, update_randoms,\
@@ -20,7 +19,6 @@ from visualiser import build_fig, draw_tstep, set_style, plot_sir
 #set seed for reproducibility
 #np.random.seed(100)
 
-# this change is for lab 7
 
 class Simulation():
     #TODO: if lockdown or otherwise stopped: destination -1 means no motion
@@ -206,7 +204,57 @@ dead: %i, of total: %i' %(self.frame, self.pop_tracker.susceptible[-1], self.pop
         plot_sir(self.Config, self.pop_tracker, size, include_fatalities,
                  title)
 
+class Environment():
+    '''
+    file that contains all functions to define destinations in the
+    environment of the simulated world.
+    '''
 
+    def build_hospital(xmin, xmax, ymin, ymax, plt, addcross=True):
+        '''builds hospital
+
+        Defines hospital and returns wall coordinates for
+        the hospital, as well as coordinates for a red cross
+        above it
+
+        Keyword arguments
+        -----------------
+        xmin : int or float
+            lower boundary on the x axis
+
+        xmax : int or float
+            upper boundary on the x axis
+
+        ymin : int or float
+            lower boundary on the y axis
+
+        ymax : int or float
+            upper boundary on the y axis
+
+        plt : matplotlib.pyplot object
+            the plot object to which to append the hospital drawing
+            if None, coordinates are returned
+
+        Returns
+        -------
+        None
+        '''
+
+        #plot walls
+        plt.plot([xmin, xmin], [ymin, ymax], color = 'black')
+        plt.plot([xmax, xmax], [ymin, ymax], color = 'black')
+        plt.plot([xmin, xmax], [ymin, ymin], color = 'black')
+        plt.plot([xmin, xmax], [ymax, ymax], color = 'black')
+
+        #plot red cross
+        if addcross:
+            xmiddle = xmin + ((xmax - xmin) / 2)
+            height = np.min([0.3, (ymax - ymin) / 5])
+            plt.plot([xmiddle, xmiddle], [ymax, ymax + height], color='red',
+                     linewidth = 3)
+            plt.plot([xmiddle - (height / 2), xmiddle + (height / 2)],
+                     [ymax + (height / 2), ymax + (height / 2)], color='red',
+                     linewidth = 3)
 
 if __name__ == '__main__':
